+++ server/bookings/queries/getBookingsByUserEmail.js
@@ -2,7 +2,7 @@ const { bookings } = require("../mockBookings");
 
 function getBookingsByUserEmail(req, res) {
   const userBookings = bookings.filter(user => user.userEmail === req.params.email);
-  userBookings ? res.json(userBookings) : res.status(404).send("Bookings were not found by user email");
+  userBookings.length ? res.json(userBookings) : res.status(404).send("Bookings were not found by user email");
 }
 
 module.exports = {
