+++ server/businesses/mutations/newBusiness.js
@@ -1,22 +1,5 @@
 const { businesses } = require("../mockBusinesses");
-const Joi = require("joi");
-
-const businessSchema = Joi.object({
-  name: Joi.string().required(),
-  description: Joi.string().required(),
-  address: Joi.string().required(),
-  category: Joi.string().required(),
-  contactPerson: Joi.string().required(),
-  email: Joi.string().email().required(),
-  images: Joi.array()
-    .items(
-      Joi.object({
-        url: Joi.string().uri().required()
-      }).required()
-    )
-    .min(1)
-    .required()
-});
+const { businessSchema } = require("../validate");
 
 function newBusiness(req, res) {
   const { error, value } = businessSchema.validate(req.body);
