+++ server/businesses/queries/getBusinessesByCategory.js
@@ -2,7 +2,7 @@ const { businesses } = require("../mockBusinesses");
 
 function getBusinessesByCategory(req, res) {
   const filteredBusinesses = businesses.filter(business => business.category.toLowerCase() === req.params.category.toLowerCase());
-  filteredBusinesses ? res.json(filteredBusinesses) : res.status(404).send("Businesses were not found by category");
+  filteredBusinesses.length ? res.json(filteredBusinesses) : res.status(404).send("Businesses were not found by category");
 }
 
 module.exports = {
