+++ server/categories/mutations/newCategory.js
@@ -1,15 +1,5 @@
 const { categories } = require("../mockCategories");
-const Joi = require("joi");
-
-const categorySchema = Joi.object({
-  name: Joi.string().required(),
-  bgcolor: Joi.object({
-    hex: Joi.string().required()
-  }).required(),
-  icon: Joi.object({
-    url: Joi.string().required()
-  }).required()
-});
+const { categorySchema } = require("../validate");
 
 function newCategory(req, res) {
   const { error, value } = categorySchema.validate(req.body);
